rename SchemaUtils to SchemaParser
