+++ src/main/antlr4/org/apache/spark/sql/ml/parser/RikaiModelSchema.g4
@@ -15,8 +15,7 @@
 grammar RikaiModelSchema;
 
 schema
-    : struct
-    | array
+    : fieldType
     ;
 
 identifier
@@ -40,8 +39,6 @@ fieldType
 field
     : name=identifier ':' fieldType   # structField
     ;
-
-
 
 STRUCT: 'struct';
 ARRAY: 'array';
